feat(openapi): create a command to print the openapi json schema
